# src/frontend/shared/openapi/wnav-openapi.yaml
openapi: "3.1.0"
info:
  title: WNAV API
  version: 1.0.0
  description: |
    作業ナビゲーションシステム REST API。terminal-api（ポート 8080）と master-api（ポート 8081）の統合仕様。
    エラーは RFC 9457 Problem Details 形式で返却する。
  contact:
    name: WNAV Engineering
    email: engineering@wnav.example.com
servers:
  - url: http://localhost:8080/api/v1
    description: terminal-api ローカル開発
  - url: http://localhost:8081/api/v1
    description: master-api ローカル開発
  - url: https://wnav-terminal.factory.local/api/v1
    description: terminal-api 本番
  - url: https://wnav-master.factory.local/api/v1
    description: master-api 本番
security:
  - bearerAuth: []
...
paths:
...
  /work-assignments:
    get:
      tags: [work-assignments]
      operationId: listWorkAssignments
      summary: 作業指示割当一覧（API-sync-005）
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: after, in: query, schema: { type: string } }
      responses:
        '200': { description: 一覧, content: { application/json: { schema: { $ref: '#/components/schemas/WorkAssignmentListEnvelope' } } } }
    post:
      tags: [work-assignments]
      operationId: createWorkAssignment
      summary: 作業指示 Push 受信（API-sync-003）
      security: []
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkAssignmentRequest' }
      responses:
        '202': { description: 受付成功, content: { application/json: { schema: { $ref: '#/components/schemas/WorkAssignmentEnvelope' } } } }
...
  /work-assignments/{id}/ack:
    post:
      tags: [work-assignments]
      operationId: ackWorkAssignment
      parameters: [ { $ref: '#/components/parameters/PathId' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: ACK 完了, content: { application/json: { schema: { $ref: '#/components/schemas/WorkAssignmentEnvelope' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 で署名された JWT を Bearer トークンとして送信する
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid, description: UUID v7 形式 }
...
  schemas:
...
    ResponseMeta:
      type: object
      required: [request_id, server_time, api_version]
      properties:
        request_id: { type: string, format: uuid }
        server_time: { type: string, format: date-time }
        api_version: { type: string, enum: [v1] }
...
    CreateWorkAssignmentRequest:
      type: object
      required: [external_order_id, external_system, work_pattern_key, target_terminal_key]
      properties:
        external_order_id: { type: string }
        external_system: { type: string }
        work_pattern_key: { type: string }
        target_terminal_key: { type: string }
        lot_id_ext: { type: string }
        suggested_worker_key: { type: string }
        suggested_equipment_key: { type: string }
        due_at: { type: string, format: date-time }
        priority: { type: integer }
    WorkAssignment:
      type: object
      properties:
        id: { type: string, format: uuid }
        assignment_id: { type: string, format: uuid }
        sop_id: { type: string, format: uuid }
        sop_name: { type: string }
        target_terminal_id: { type: string, format: uuid }
        lot_id: { type: string, format: uuid, nullable: true }
        lot_number: { type: string, nullable: true }
        suggested_worker_id: { type: string, format: uuid, nullable: true }
        suggested_equipment_id: { type: string, format: uuid, nullable: true }
        due_at: { type: string, format: date-time, nullable: true }
        priority: { type: integer }
        status: { type: string }
        received_at: { type: string, format: date-time }
        acknowledged_at: { type: string, format: date-time, nullable: true }
    WorkAssignmentEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/WorkAssignment' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    WorkAssignmentListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/WorkAssignment' } }
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - type: object
              properties:
                limit: { type: integer }
                has_more: { type: boolean }
                next_cursor: { type: string, nullable: true }
...